==> training-with-artificial-data/01_data_prep.ipynb <==
# Prepare data for training dataframe

The aim is to produce a training table where each `order_id` is mapped to all coupons available in the departments from which products were selected in that order on the date the order was made, along with information (`True`/`False`) if such available coupon was used in that order.

The steps are as follows:
1. Create `coupon_date_department` table which maps coupons to applicable departments, to coupon validity dates
2. Create `order_date_department` table which maps orders to departments from which products were ordered, to order dates.
3. Create `order_coupons_available` by merging the tables above on date and department - this will create a table which maps orders and coupons available at that order date, but limited by departments (i.e. rows are selected only if order contained products from a department for which there was a coupon)
4. Create `order_coupons_used` - by dropping columns from `order_details`, leaving only a mapping between order and coupons used in that order.
5. Cobine 3 and 4 to create a dataset with coupons and orders mapped, with information whether coupon was used or not.
6. Add details about customers (such as age, gender, etc., and also some statistics about customer purchases)
7. Add details about coupons (type, discount, mean product pric, etc.)

import datetime
import os

from IPython.display import Image
import numpy as np
import pandas as pd

Image('../data_diagram.png')

## Read and Prepare Tables

data_dir = 'data_0419_0'

coupons = pd.read_csv(os.path.join(data_dir, 'coupons.csv'))
coupons.info()

coupons.sample(5)

coupons.rename(columns={'id': 'coupon_id'}, inplace=True)
coupons.start_date = pd.to_datetime(coupons.start_date, format='%Y-%m-%d')
coupons.end_date = pd.to_datetime(coupons.end_date, format='%Y-%m-%d')

coupon_product = pd.read_csv(os.path.join(data_dir, 'coupon_product.csv'))
coupon_product.info()

products = pd.read_csv(os.path.join(data_dir, 'products.csv'))
products.info()

products.sample(5)

products.rename(columns={'id': 'product_id'}, inplace=True)

orders = pd.read_csv(os.path.join(data_dir, 'orders.csv'))
orders.info()

orders.rename(columns={'id': 'order_id', 'order_date': 'date'}, inplace=True)
orders.date = pd.to_datetime(orders.date, format='%Y-%m-%d').dt.date

order_details = pd.read_csv(os.path.join(data_dir, 'order_details.csv'))
order_details.info()

order_details.drop(['id'], axis=1, inplace=True)

order_details.sample(5)

order_details.coupon_id.isnull().sum() / len(order_details)

customers = pd.read_csv(os.path.join(data_dir, 'customers.csv'))
customers.info()

customers.rename(columns={'id': 'customer_id'}, inplace=True)

customers.sample(5)

## Data Prep

#### Step 1. Create `coupon_date_department` table

1. Create `coupon_department` mapping coupons to departments for which they are valid
2. Create `coupon_dates` mapping coupons to all dates on which they were valid
3. Merge the two into `coupon_department_dates`

# Step 1.1 Map coupons to departments
coupon_department = pd.merge(coupon_product, products[['product_id', 'department']], on='product_id')\
                      .drop(['product_id'], axis=1).drop_duplicates()

# Coupon_product does not include coupons valid for all products in a department, add this info here
department_coupons = coupons.loc[coupons.type == 'department'][['coupon_id', 'department']]
coupon_department = coupon_department.append(department_coupons).sort_values(by='coupon_id').reset_index(drop=True)
coupon_department

# Validate that all coupons are present
assert len(coupon_department) == len(coupons.coupon_id.unique())

# Step 1.2 Map coupons to dates on which they were valid
coupon_dates = coupons.drop(['type', 'department', 'discount', 'how_many'], axis=1)
# Get the earliest and latest date in the dataset
start = coupon_dates.start_date.min()
end = coupon_dates.end_date.max()
days = (end-start).days + 1
print(f'Start: {start.date()}, end: {end.date()}. {days} days')

# Create a dataframe with row for each day from the earliest to the latest date in the set
all_dates = pd.DataFrame(pd.date_range(start=start, end=end, freq='D'), columns=['date'])
assert days == len(all_dates)

# Step 1.2.1: Perform a cross join of `all_dates` and `coupon_dates` - which contains info on validity periods
coupon_dates['key'] = 1
all_dates['key'] = 1
coupon_dates = pd.merge(coupon_dates, all_dates, on='key').drop('key', axis=1)

# Step 1.2.2 Drop rows where a date does not fall within the validity period of a coupon
coupon_dates = coupon_dates[(coupon_dates['date'] >= coupon_dates['start_date']) & \
                            (coupon_dates['date'] <= coupon_dates['end_date'])]
coupon_dates.drop(['start_date', 'end_date'], axis=1, inplace=True)
coupon_dates

# Validate coupon_dates is consistent with the original data in terms of coupon validity dates
coupons['days_valid'] = (coupons.end_date - coupons.start_date).dt.days + 1
df = pd.merge(coupons, coupon_dates.groupby(by='coupon_id').count().rename(columns={'date': 'days_valid'}), on='coupon_id')
assert 0 == len(df.loc[df.days_valid_x != df.days_valid_y])
coupons.drop(['days_valid'], axis=1, inplace=True)

# Validate no coupon has been lost
assert len(coupons) == len(coupon_dates.coupon_id.unique())

coupon_dates.groupby('date').count().median()

# Step 1.3 Merge coupon_department and coupon_dates
coupon_date_department = pd.merge(coupon_dates, coupon_department, on='coupon_id', how='left')
coupon_date_department.date = coupon_date_department.date.dt.date
coupon_date_department

#### Step 2. Create `order_date_department` table

1. Map `order_id` and product `department` -> `order_department`
2. Add dates by joining with `orders` -> `order_date_department`

order_department = pd.merge(order_details[['order_id', 'product_id']], products[['product_id', 'department']],
                            on='product_id', how='left').drop(['product_id'], axis=1).drop_duplicates()
order_department

order_date_department = pd.merge(order_department, orders[['order_id', 'date']], on='order_id', how='right')
order_date_department

#### Step 3. Create `order_coupons_available`

order_coupons_available = pd.merge(order_date_department, coupon_date_department, on=['date', 'department'], how='left')\
                            .drop(['department', 'date'], axis=1).drop_duplicates()
order_coupons_available

order_coupons_available.groupby('order_id').count().sort_values(by='coupon_id', ascending=False)

#### Step 4. Create `order_coupons_used`

order_coupon_used = order_details[['order_id', 'coupon_id']].dropna().drop_duplicates()
order_coupon_used

#### Step 5. Combine `order_coupons_available` and `order_coupons_used` to get final info on which coupons were used and which, although they were available, were ignored

order_coupon_used.set_index(['order_id', 'coupon_id'], inplace=True)
order_coupons_available.set_index(['order_id', 'coupon_id'], inplace=True)

order_coupons_available['coupon_used'] = order_coupons_available.index.isin(order_coupon_used.index)

order_coupons = order_coupons_available.reset_index().dropna().reset_index(drop=True)

order_coupons.coupon_used.value_counts(normalize=True)

## Prepare customer data
- From `customers` table, get:
    - gender
    - age bracket: young < 30, medium >= 30 & < 60, old >= 60
- From `order_details` get:
    - sum of unique products bought
    - sum of unique products bought at a discount
    - sum total of coupons used
    - total products bought
    - mean price paid
    - mean discount used

customer_demo = customers.drop(['name', 'phone', 'address', 'city', 'state', 'postalCode', 'country', 'creditLimit'],
                               axis=1)

customer_demo['age_bracket'] = None
customer_demo.loc[customer_demo.age < 30, 'age_bracket'] = 'young'
customer_demo.loc[(customer_demo.age >= 30) & (customer_demo.age < 60), 'age_bracket'] = 'mid'
customer_demo.loc[(customer_demo.age >= 60), 'age_bracket'] = 'old'
customer_demo.age = customer_demo.age_bracket
customer_demo.drop(['age_bracket'], axis=1, inplace=True)

customer_demo.info()

customer_demo.rename(columns={'age': 'cust_age', 'gender': 'cust_gender'}, inplace=True)

cust_orders = pd.merge(order_details, orders[['order_id', 'customer_id']], on='order_id', how='left')\
                .drop(['order_id'], axis=1)
cust_orders['discount'] = 100 * ((cust_orders.original_price - cust_orders.buy_price) / cust_orders.original_price)
cust_orders.drop(['original_price'], axis=1, inplace=True)
cust_orders.sample(5)

cust_stats = pd.pivot_table(cust_orders,
                            values=['product_id', 'buy_price', 'coupon_id', 'discount'],
                            index='customer_id',
                            aggfunc={
                                'product_id': lambda x: len(set(x)),  # sum of unique products bought
                                'coupon_id': lambda x: x.notnull().sum(),  # total coupons used
                                'discount': lambda x: np.round(np.mean(x), decimals=2),  # mean discount used
                                'buy_price': lambda x: np.round(np.mean(x), decimals=2)  # mean price paid
                            })
cust_stats.rename(columns={
    'product_id': 'cust_unique_products',
    'coupon_id': 'cust_total_coupons',
    'discount': 'cust_mean_discount',
    'buy_price': 'cust_mean_buy_price'
}, inplace=True)

cust_stats['cust_unique_products_coupon'] = cust_orders.loc[cust_orders.coupon_id.notnull()]\
    .groupby('customer_id').agg({'product_id': 'nunique'})
cust_stats.fillna(value=0, inplace=True)

cust_stats['cust_total_products'] = cust_orders.groupby('customer_id').count().product_id

customer_data = pd.merge(customer_demo, cust_stats, on='customer_id', how='left')
customer_data

## Prepare coupon data
- From `coupons` table, take:
    - type
    - discount
    - department
- From merging `products` and `coupon_product`, `coupons`, take:
    - mean_product_price
    - no_prods_available

coupon_info = coupons.drop(['start_date', 'end_date'], axis=1)
coupon_info.loc[coupon_info.how_many == -1, 'how_many'] = 1
coupon_info.info()

coupon_info.rename(
    columns={
        'type': 'coupon_type',
        'how_many': 'coupon_how_many',
        'discount': 'coupon_discount',
        'department': 'coupon_dpt'},
    inplace=True)

coupon_stats = pd.merge(coupon_product, products[['product_id', 'buy_price']], on='product_id').drop_duplicates()

# Coupon_product does not include coupons valid for all products in a department, add those product prices here
department_coupons = coupons.loc[coupons.type == 'department'][['coupon_id', 'department']]\
                            .merge(products[['department', 'buy_price', 'product_id']], on='department', how='left')\
                            .drop(['department'], axis=1)
coupon_stats = coupon_stats.append(department_coupons).reset_index(drop=True)

coupon_stats = pd.pivot_table(coupon_stats, index='coupon_id', values=['buy_price', 'product_id'],
                              aggfunc={
                                  'buy_price': lambda x: np.round(np.mean(x), decimals=2),
                                  'product_id': lambda x: len(set(x))
                              })
coupon_stats.rename(columns={'buy_price': 'coupon_mean_prod_price', 'product_id': 'coupon_prods_avail'}, inplace=True)
coupon_stats

coupon_data = pd.merge(coupon_info, coupon_stats, on='coupon_id', how='left')
coupon_data

## Merge everyghing into one dataframe

1. Add `customer_id` to `order_coupons` (from `orders`) -> `final`
2. Merge `final` with `customer_data`
3. Merge `final` with `coupon_data`

final = pd.merge(order_coupons, orders[['order_id', 'customer_id']], on='order_id', how='left')
final = pd.merge(final, customer_data, on='customer_id', how='left')
final = pd.merge(final, coupon_data, on='coupon_id', how='left')
final

final.coupon_used.value_counts(normalize=True)

## Droping ids, encoding

train = final.drop(['order_id', 'coupon_id', 'customer_id'], axis=1)

train.coupon_used = train.coupon_used.astype(int)
train = pd.get_dummies(train, columns=['cust_gender', 'cust_age', 'coupon_type', 'coupon_dpt'])

train.columns

## Save as csv

train.to_csv(os.path.join(data_dir, 'train.csv'), index=False)

final.drop(['order_id'], axis=1).to_csv(os.path.join(data_dir, 'train_before_encoding.csv'), index=False)
