==> training-with-artificial-data/02_training_automl.ipynb <==
import os

import h2o
from h2o.automl import H2OAutoML

h2o.init()

data_dir = 'data_0419_0'
train_file_path = os.path.join(data_dir, 'train.csv')

train_df = h2o.import_file(train_file_path)
train_df.describe()

y = 'coupon_used'
x = train_df.columns
x.remove(y)
train_df['coupon_used'] = train_df['coupon_used'].asfactor()

aml = H2OAutoML(max_models=10, seed=1, balance_classes=True)
aml.train(x=x, y=y, training_frame=train_df)

aml.leaderboard

h2o.get_model('GBM_4_AutoML_20210426_102535')
