==> training-with-artificial-data/04_demo_data_selection.ipynb <==
import os
import pickle

import numpy as np
import pandas as pd
from scipy import stats

data_dir = 'data_0419_0'

DPT_BOYS = 'Boys'
DPT_GIRLS = 'Girls'
DPT_MEN = 'Men'
DPT_SPORT = 'Sport'
DPT_WOMEN = 'Women'
DEPARTMENTS = [DPT_BOYS, DPT_GIRLS, DPT_MEN, DPT_SPORT, DPT_WOMEN]

COUPON_TYPES = ['buy_all', 'department', 'buy_more', 'just_discount']

customers = pd.read_csv(os.path.join(data_dir, 'customers.csv'))
customers.info()

customers.rename(columns={'id': 'customer_id'}, inplace=True)

### Let the model calculate coupon redemption probabilities

with open(os.path.join(data_dir, 'pickled_model_gbm_smote'), 'rb') as f:
    model = pickle.load(f)

df_probs_full = pd.read_csv(os.path.join(data_dir, 'train_before_encoding.csv')).drop(['coupon_used'], axis=1)

encoded_df = pd.get_dummies(df_probs_full, columns=['cust_gender', 'cust_age', 'coupon_type', 'coupon_dpt'])\
            .drop(['customer_id', 'coupon_id'], axis=1)
df_probs_full['probs'] = model.predict_proba(encoded_df)[:, 1]
df_probs_full

df_probs = df_probs_full[['coupon_id', 'coupon_dpt', 'coupon_type', 'customer_id', 'probs']].drop_duplicates()
df_probs

### Find coupons which have various scores for various customers

high = df_probs.loc[df_probs.probs >= 0.5]
mid = df_probs.loc[(df_probs.probs < 0.5) & (df_probs.probs >= 0.3)]
low = df_probs.loc[df_probs.probs < 0.3]

coupons = high[['coupon_id', 'coupon_dpt', 'coupon_type']].merge(mid['coupon_id'], on='coupon_id', how='inner')\
    .merge(low['coupon_id'], on='coupon_id', how='inner')\
    .drop_duplicates()
coupons.coupon_dpt.value_counts()

### Split data into groups of 3 coupons per dpt (daily set of coupons)

coupons.sample(frac=1)
coupon_set_0 = coupons.groupby('coupon_dpt').apply(lambda x: x.iloc[:3]).reset_index(drop=True)
coupon_set_1 = coupons.groupby('coupon_dpt').apply(lambda x: x.iloc[3:6]).reset_index(drop=True)
coupon_set_2 = coupons.groupby('coupon_dpt').apply(lambda x: x.iloc[6:9]).reset_index(drop=True)

# Fill in with random coupons if a set has less than 3 coupons per dpt
coupons_to_fill_in = df_probs[['coupon_dpt', 'coupon_id', 'coupon_type']][~df_probs.coupon_id.isin(coupons['coupon_id'])]
def fill_in_missing_coupons(coupon_set):
    for dpt_name in DEPARTMENTS:
        missing = 3 - coupon_set.coupon_dpt.value_counts().get(dpt_name, 0)
        rows_to_add = coupons_to_fill_in.loc[coupons_to_fill_in.coupon_dpt == dpt_name].sample(missing)
        coupon_set = coupon_set.append(rows_to_add)
    return coupon_set
coupon_set_0 = fill_in_missing_coupons(coupon_set_0)
coupon_set_1 = fill_in_missing_coupons(coupon_set_1)
coupon_set_2 = fill_in_missing_coupons(coupon_set_2)

### Select customers who have high scores within a department

def select_customers(coupon_set):
    df = pd.merge(coupon_set[['coupon_id']], df_probs, on='coupon_id', how='left')
    customers = pd.DataFrame(columns=['customer_id', 'coupon_dpt'])
    for dpt_name in DEPARTMENTS:
        dpt_customers = df.loc[df.coupon_dpt == dpt_name]\
            .sort_values('probs', ascending=False)[['customer_id']].drop_duplicates().head(5)
        dpt_customers['coupon_dpt'] = dpt_name
        customers = customers.append(dpt_customers)
    return customers

customers_0 = select_customers(coupon_set_0)
customers_1 = select_customers(coupon_set_1)
customers_2 = select_customers(coupon_set_2)

### Combine the coupons and customers

def create_data_set(customers, coupons):
    data = pd.merge(customers['customer_id'], df_probs_full, on='customer_id', how='left')\
        .merge(coupons['coupon_id'], on='coupon_id', how='right').drop_duplicates()\
        .sort_values(['coupon_dpt', 'customer_id', 'coupon_id'])
    return data

data_set_0 = create_data_set(customers_0, coupon_set_0)
data_set_1 = create_data_set(customers_1, coupon_set_1)
data_set_2 = create_data_set(customers_2, coupon_set_2)
data_set_0

### Save data to csv files
- `customer_stats.csv` - additional statistical data which are missing from original data and are required as input to the model
- `coupon_stats.csv` - additional coupon data
- `good_customer_coupon_pairs.csv` - customers for which the model returns many coupon 'hits', paired with the 'hit' coupons
- `good_customer_id_coupon_data.csv` - the same set of customer-coupon pairs, but instead of coupon_ids, coupon stats are present
- `good_customer_coupon_full.csv` - 'good' customer-coupon pairs along with all the data required as input to the model

demo_data_dir = os.path.join(data_dir, 'demo_data')
if not os.path.exists(demo_data_dir):
    os.makedirs(demo_data_dir)

#### Coupon stats

for i, ds in enumerate([data_set_0, data_set_1, data_set_2]):
    coupon_stats = ds[['coupon_id', 'coupon_type', 'coupon_dpt', 'coupon_discount', 'coupon_how_many',
                       'coupon_mean_prod_price', 'coupon_prods_avail']].drop_duplicates()\
        .sort_values(by='coupon_id')
    coupon_stats.to_csv(os.path.join(demo_data_dir, f'coupon_stats_{i}.csv'), index=False)

#### Customer stats

for i, ds in enumerate([data_set_0, data_set_1, data_set_2]):
    customer_stats = ds[['customer_id', 'cust_gender', 'cust_age', 'cust_mean_buy_price', 'cust_total_coupons',
                         'cust_mean_discount', 'cust_unique_products', 'cust_unique_products_coupon',
                         'cust_total_products']].drop_duplicates()\
        .sort_values(by='customer_id')
    customer_stats.to_csv(os.path.join(demo_data_dir, f'customer_stats_{i}.csv'), index=False)

#### Full data - with probabilities

for i, ds in enumerate([data_set_0, data_set_1, data_set_2]):
    ds.to_csv(os.path.join(demo_data_dir, f'full_data_set_{i}.csv'), index=False)
