==> training-with-artificial-data/05_data_4_cache_db.ipynb <==
# Generate data for cache db
This notebook can be used to generate data for Recommendation Service Cache DB

import os

import numpy as np
import pandas as pd

## Read data

data_dir = 'data_0419_0'

db_data_dir = os.path.join(data_dir, 'data4db')
if not os.path.exists(db_data_dir):
    os.makedirs(db_data_dir)

### product_info table

```SQL
CREATE TABLE product_info (
    product_id INT,
    name VARCHAR(256),
    category VARCHAR(50),
    sizes VARCHAR(50),
    vendor VARCHAR(50),
    description VARCHAR(256),
    buy_price REAL,
    department VARCHAR(10),
    PRIMARY KEY (product_id)
);

COPY product_info FROM '<<DATA_PATH>>/products.csv' DELIMITER ',' CSV HEADER;
```

products = pd.read_csv(os.path.join(data_dir, 'products.csv'))
products.sample()

products.rename(columns={'id': 'product_id'}, inplace=True)

products.to_csv(os.path.join(db_data_dir, 'products.csv'), index=False)

### coupon_info table

```SQL
CREATE TABLE coupon_info (
  coupon_id INT,
  coupon_type REAL,
  department VARCHAR(10),
  discount INT,
  how_many_products_required INT,
  product_mean_price REAL,
  products_available INT,
  start_date VARCHAR(10),
  end_date VARCHAR(10),
  PRIMARY KEY (coupon_id)
);

COPY coupon_info FROM '<<DATA_PATH>>/coupon_info.csv' DELIMITER ',' CSV HEADER;
```

coupons = pd.read_csv(os.path.join(data_dir, 'coupons.csv'))
coupons.sample()

coupons.rename(columns={'id': 'coupon_id', 'type': 'coupon_type', 'how_many': 'how_many_products_required'}, inplace=True)
coupons.start_date = pd.to_datetime(coupons.start_date, format='%Y-%m-%d')
coupons.end_date = pd.to_datetime(coupons.end_date, format='%Y-%m-%d')

coupon_product = pd.read_csv(os.path.join(data_dir, 'coupon_product.csv'))
coupon_product.info()

# In 'department'-type coupons, change 'how_many_products_required' from -1 to 1
coupons.loc[coupons.coupon_type == 'department', 'how_many_products_required'] = 1

# coupon_product table is missing mapping for 'department'-type coupons, fill that in
dept_coupon_prod = pd.merge(
    coupons.loc[coupons.coupon_type == 'department'][['coupon_id', 'department']],
    products[['product_id', 'department']],
    on='department',
    how='left').drop('department', axis=1)

coupon_product = coupon_product.append(dept_coupon_prod)
coupon_product

coupon_product_price = pd.merge(coupon_product, coupons[['coupon_id']], on='coupon_id', how='left')\
    .merge(products[['product_id', 'buy_price']], on='product_id', how='left').drop_duplicates()
coupon_product_price

coupon_stats = pd.pivot_table(coupon_product_price, values=['product_id', 'buy_price'], index='coupon_id',
                              aggfunc={
                                  'product_id': len,
                                  'buy_price': np.mean
                              })
coupon_stats.buy_price = coupon_stats.buy_price.round(2)
coupon_stats.rename(columns={'buy_price': 'product_mean_price', 'product_id': 'products_available'}, inplace=True)

coupon_info = pd.merge(coupons, coupon_stats, on='coupon_id', how='left')
coupon_info

coupon_info.to_csv(os.path.join(db_data_dir, 'coupon_info.csv'), index=False)

### coupon_product table

```SQL
CREATE TABLE coupon_product (
    coupon_id INT,
    product_id INT,
    FOREIGN KEY (coupon_id) REFERENCES coupon_info(coupon_id),
    FOREIGN KEY (product_id) REFERENCES products(product_id)
);

COPY coupon_product FROM '<<DATA_PATH>>/coupon_product.csv' DELIMITER ',' CSV HEADER;
```

coupon_product.drop_duplicates().to_csv(os.path.join(db_data_dir, 'coupon_product.csv'), index=False)

### customer_info table

```SQL
CREATE TABLE customer_info (
  customer_id INT,
  gender VARCHAR(1),
  age INT,
  mean_buy_price REAL,
  unique_products_bought INT,
  unique_products_bought_with_coupons: INT,
  total_items_bought: INT,
  mean_discount_received: REAL,
  total_coupons_used: INT
  PRIMARY KEY (customer_id)
);

COPY customer_info FROM '<<DATA_PATH>>/customer_info.csv' DELIMITER ',' CSV HEADER;
```

customers = pd.read_csv(os.path.join(data_dir, 'customers.csv'))
customers.info()

customers = customers[['id', 'gender', 'age']].rename(columns={'id': 'customer_id'})
customers

orders = pd.read_csv(os.path.join(data_dir, 'orders.csv'))
orders.info()

orders.rename(columns={'id': 'order_id'}, inplace=True)

order_details = pd.read_csv(os.path.join(data_dir, 'order_details.csv'))
order_details.info()

order_details.drop('id', axis=1, inplace=True)

cust_orders = pd.merge(orders, order_details, on='order_id', how='left')
cust_orders['discount'] = 100 * ((cust_orders.original_price - cust_orders.buy_price) / cust_orders.original_price)

cust_stats = pd.pivot_table(cust_orders,
                            values=['product_id', 'buy_price', 'coupon_id', 'discount'],
                            index='customer_id',
                            aggfunc={
                                'product_id': lambda x: len(set(x)),  # sum of unique products bought
                                'coupon_id': lambda x: x.notnull().sum(),  # total coupons used
                                'discount': lambda x: np.round(np.mean(x), decimals=2),  # mean discount used
                                'buy_price': lambda x: np.round(np.mean(x), decimals=2)  # mean price paid
                            })
cust_stats.rename(columns={
    'product_id': 'unique_products_bought',
    'coupon_id': 'total_coupons_used',
    'discount': 'mean_discount_received',
    'buy_price': 'mean_buy_price'
}, inplace=True)

cust_stats['unique_products_bought_with_coupons'] = cust_orders.loc[cust_orders.coupon_id.notnull()]\
    .groupby('customer_id').agg({'product_id': 'nunique'})
cust_stats.fillna(value=0, inplace=True)

cust_stats['total_items_bought'] = cust_orders.groupby('customer_id').count().product_id

customer_info = pd.merge(customers[['customer_id', 'gender', 'age']], cust_stats, on='customer_id', how='left')
customer_info

customer_info.dropna(inplace=True)
customer_info.total_coupons_used = customer_info.total_coupons_used.astype(int)
customer_info.unique_products_bought = customer_info.unique_products_bought.astype(int)
customer_info.unique_products_bought_with_coupons = customer_info.unique_products_bought_with_coupons.astype(int)
customer_info.total_items_bought = customer_info.total_items_bought.astype(int)

customer_info.to_csv(os.path.join(db_data_dir, 'customer_info.csv'), index=False)
